Applied SOLID principal rules in freshness.go file and cassandra.go file. Also modified the freshness.md file
